Post-integration code cleanup from Perforce mainline
- whitespace
- fix an x64 compilation warning

+++ Code/Core/FileIO/PathUtils.cpp
@@ -187,10 +187,10 @@
 //------------------------------------------------------------------------------
 /*static*/ void PathUtils::StripFileExtension( AString & filePath )
 {
-    const char * lastDot = filePath.FindLast('.');
+    const char * lastDot = filePath.FindLast( '.' );
     if (lastDot)
     {
-        filePath.SetLength(lastDot - filePath.Get());
+        filePath.SetLength( (uint32_t)( lastDot - filePath.Get() ) );
     }
 }
